requirements.txt
numpy
pandas
scikit-learn
opencv-python
matplotlib
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.6
pillow
